refactor: restructure operator public page
