+++ src/pages-layer/operator-public/ui/VerifiedBadge/VerifiedBadge.tsx
